requirements.txt
Django==4.0.2
pre-commit==2.17.0
flake8==4.0.1
flake8-isort==4.1.1
black==22.1.0
pylint-django==2.5.2
djangorestframework==3.13.1
drf-nested-routers==0.93.4
djangorestframework-simplejwt==5.1.0
